small fixes (AddLoan: InterestRate now has a value Model: InteresRate is now double instead of float)

+++ BACKEND/BACKEND/Model/Loan.cs
@@ -3,7 +3,7 @@
     public class Loan
     {
         public int Years { get; set; }
-        public float InterestRate { get; set; }
+        public double InterestRate { get; set; }
         public int RepayAmount { get; set; }
     }
 }
